Made spawn blocks functional
Character now spawns where the spawn block is placed on the level.
If no spawn block is found, then the character spawns at default (30,
30)
If more than 1 spawn block is present, it will use the highest spawn
block, and if there are multiple on the top row it will use the
leftmost.