Rewrite of UI Code
Removed the fraction method (was not appropriate for drawing the lines).
Line functionality is now implemented, using UI.drawLine(xStart, yStart,
xFinish, yFinish)). This draws a line onto the screen (without the
graphics library currently in place).

The aim is to replace the graphics library currently in place, so that
fonts can be rendered in front of the camera view's (currently not an
option).